src/requirements.txt
black
anyio
cffi
click
cryptography
fastapi
flake8
h11
idna
isort
pycparser
pydantic
PyJWT
python-dotenv
secure
sniffio
starlette
typing_extensions
uvicorn
authlib 
flask-jwt-extended
fastapi
pydantic
uvicorn
pipenv
couchbase
websockets
prometheus_client
python-dotenv
starlette
starlette_prometheus